仓库类 BranchState 改为 BranchNodesState；新建数据类 BranchNodeState（神 TM 命名）
