+++ dev-doc/词汇约定.md
@@ -56,37 +56,39 @@ IF_ICMPLE L4 对应 n>1919
 
 IFEQ L4 对应 list.contains(obj)
 
-它们是**字节码分支结点**。
+我们称它们是**字节码分支结点**。
 
 （字节码里爱放“不成立”的情况……“成立则跳转，不成立则继续”）
 
 
 
+16 个“比较后跳转”，5 个“比较后记录结果”
+
 （下表：同一个 Arity 分组里按 Opcode 排序 asc）
 
-|                              | Arity | 功能                                     | JVM Instruction | Trace 语 |
-| ---------------------------- | ----- | ---------------------------------------- | --------------- | -------- |
-| 字节码分支结点（比较后跳转） | 二元  | 整数比较                                 | IF_ICMPEQ       | I$==     |
-|                              |       |                                          | IF_ICMPNE       | I$!=     |
-|                              |       |                                          | IF_ICMPLT       | I$<      |
-|                              |       |                                          | IF_ICMPGE       | I$>=     |
-|                              |       |                                          | IF_ICMPGT       | I$>      |
-|                              |       |                                          | IF_ICMPLE       | I$<=     |
-|                              |       | 地址判等                                 | IF_ACMPEQ       | A$==     |
-|                              |       |                                          | IF_ACMPNE       | A$!=     |
-|                              | 一元  | 与 0 比较                                | IFEQ            | IZ$==    |
-|                              |       |                                          | IFNE            | IZ$!=    |
-|                              |       |                                          | IFLT            | IZ$<     |
-|                              |       |                                          | IFGE            | IZ$>=    |
-|                              |       |                                          | IFGT            | IZ$>     |
-|                              |       |                                          | IFLE            | IZ$<=    |
-|                              |       | 与 null 判等                             | IFNULL          | N$==     |
-|                              |       |                                          | IFNONNULL       | N$!=     |
-| 比较后记录结果               | 二元  | 比较 long。push 0 if ==, 1 if >, -1 if < | LCMP            | L$CMP    |
-|                              |       | 同上 but 比较 double。若有 NaN 则 1      | DCMPG           | D$CMPG   |
-|                              |       | 同上 but 若有 NaN 则 -1                  | DCMPL           | D$CMPL   |
-|                              |       | 同上上 but 比较 float                    | FCMPG           | F$CMPG   |
-|                              |       | 同上上 but 比较 float                    | FCMPL           | F$CMPL   |
+| 比较之后干嘛？ | Arity | 功能                                       | JVM Instruction | Trace 语 |
+| -------------- | ----- | ------------------------------------------ | --------------- | -------- |
+| 比较后跳转     | 二元  | 整数比较                                   | IF_ICMPEQ       | I$==     |
+|                |       |                                            | IF_ICMPNE       | I$!=     |
+|                |       |                                            | IF_ICMPLT       | I$<      |
+|                |       |                                            | IF_ICMPGE       | I$>=     |
+|                |       |                                            | IF_ICMPGT       | I$>      |
+|                |       |                                            | IF_ICMPLE       | I$<=     |
+|                |       | 地址判等                                   | IF_ACMPEQ       | A$==     |
+|                |       |                                            | IF_ACMPNE       | A$!=     |
+|                | 一元  | 与 0 比较                                  | IFEQ            | IZ$==    |
+|                |       |                                            | IFNE            | IZ$!=    |
+|                |       |                                            | IFLT            | IZ$<     |
+|                |       |                                            | IFGE            | IZ$>=    |
+|                |       |                                            | IFGT            | IZ$>     |
+|                |       |                                            | IFLE            | IZ$<=    |
+|                |       | 与 null 判等                               | IFNULL          | N$==     |
+|                |       |                                            | IFNONNULL       | N$!=     |
+| 比较后记录结果 | 二元  | 比较俩 long。push 0 if ==, 1 if >, -1 if < | LCMP            | L$CMP    |
+|                |       | 同上 but 比较 double。若有 NaN 则 1        | DCMPG           | D$CMPG   |
+|                |       | 同上 but 若有 NaN 则 -1                    | DCMPL           | D$CMPL   |
+|                |       | 同上上 but 比较 float                      | FCMPG           | F$CMPG   |
+|                |       | 同上上 but 比较 float                      | FCMPL           | F$CMPL   |
 
 参考：
