--- test_agent_trace_reader/src/main/java/cn/yyx/labtask/runtime/memory/state/BranchState.java
@@ -1,27 +1,0 @@
-package cn.yyx.labtask.runtime.memory.state;
-
-import java.util.Map;
-import java.util.TreeMap;
-
-public class BranchState {
-
-  Map<String, Integer> branch_state = new TreeMap<String, Integer>();
-
-  public BranchState() {}
-
-  public Integer GetBranchState(String sig) {
-    return branch_state.get(sig);
-  }
-
-  public void PutBranchState(String sig, Integer state) {
-    branch_state.put(sig, state);
-  }
-
-  public boolean BranchHasBeenIteratedOver(String sig) {
-    Integer state = branch_state.get(sig);
-    if (state != null && state == 0) {
-      return true;
-    }
-    return false;
-  }
-}
