+++ test_agent_trace_reader/src/main/java/cn/yyx/labtask/test_agent_trace_reader/TracePairComparator.java
@@ -1,24 +1,383 @@
 package cn.yyx.labtask.test_agent_trace_reader;
 
-import java.util.Map;
+import cn.yyx.labtask.runtime.memory.state.BranchNodesState;
+import cn.yyx.labtask.runtime.round.testgen.TestModel;
+
+import java.util.*;
 
 /**
- * 比较两个 Trace——这两个 trace 是一个 sequence 变异前和变异后的版本——产生 reward。
+ * 比较两个 Trace——这两个 trace 应是一个 sequence 变异前和变异后的版本分别执行产生的——产生 reward。
  *
  * <p>可以产生
  */
 public class TracePairComparator {
 
+  TestModel model = null;
 
   /**
    * 大问题是 before after 里的 key 集不一样时怎么办 :(
+   *
+   * <p>大概要用 this class 存一些"已覆盖分支"等状态？
    *
    * @param before
    * @param after
    * @return
    */
-  public static Map<String, Double> compare(
+  public Map<String, Double> compare(
       Map<String, ValuesOfBranch> before, Map<String, ValuesOfBranch> after) {
+    Set<String> commonKeys = new TreeSet<>(before.keySet()); // copy, for the
+    commonKeys.retainAll(after.keySet()); // in-place sets intersection.
+
+    TreeMap<String, Double> result = new TreeMap<>();
+    for (String b : commonKeys) {}
+
     return null;
+  }
+
+  /**
+   * YYX Old wisdom...
+   *
+   * @param previous_branch_signature
+   * @param current_branch_signature
+   * @return
+   */
+  private Map<String, Integer> BuildGuidedModel(
+      Map<String, ValuesOfBranch> previous_branch_signature,
+      Map<String, ValuesOfBranch> current_branch_signature) {
+    Map<String, Integer> influence = new TreeMap<String, Integer>();
+
+    BranchNodesState branch_state = model.GetState();
+
+    Set<String> pset = previous_branch_signature.keySet();
+    Iterator<String> pitr = pset.iterator();
+    while (pitr.hasNext()) {
+      String sig = pitr.next();
+      if (branch_state.BranchHasBeenIteratedOver(sig)) {
+        continue;
+      }
+      ValuesOfBranch previous_vob = previous_branch_signature.get(sig);
+      ValuesOfBranch vob = current_branch_signature.get(sig);
+      influence.put(sig, -1);
+      if (vob != null) {
+        switch (vob.GetCmpOptr()) {
+          case "D$CMPG":
+          case "D$CMPL":
+          case "F$CMPG":
+          case "F$CMPL":
+          case "L$CMP":
+            {
+              Integer state = branch_state.GetBranchState(sig);
+              if (state == null) {
+                state = 0b111;
+              }
+              double prev_v1 = previous_vob.GetBranchValue1();
+              double prev_v2 = previous_vob.GetBranchValue2();
+
+              double v1 = vob.GetBranchValue1();
+              double v2 = vob.GetBranchValue2();
+              if (v1 == v2) {
+                state &= 0b101;
+              } else {
+                if (v1 > v2) {
+                  state &= 0b110;
+                } else {
+                  state &= 0b011;
+                }
+              }
+              branch_state.PutBranchState(sig, state);
+              int state_copy = state;
+              int position = 1;
+              while (state_copy > 0) {
+                int bit = state_copy & 0b1;
+                if (bit == 1) {
+                  switch (position) {
+                    case 1:
+                      {
+                        double gap = (v1 - v2) - (prev_v1 - prev_v2);
+                        if (gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 2:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 3:
+                      {
+                        double gap = (v1 - v2) - (prev_v1 - prev_v2);
+                        if (gap < 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                  }
+                }
+                state_copy >>= 1;
+                position++;
+              }
+            }
+            break;
+          case "I$==":
+          case "I$!=":
+          case "A$==":
+          case "A$!=":
+          case "IZ$==":
+          case "IZ$!=":
+          case "N$!=":
+          case "N$==":
+            {
+              Integer state = branch_state.GetBranchState(sig);
+              if (state == null) {
+                state = 0b11;
+              }
+              double prev_v1 = previous_vob.GetBranchValue1();
+              double prev_v2 = previous_vob.GetBranchValue2();
+
+              double v1 = vob.GetBranchValue1();
+              double v2 = vob.GetBranchValue2();
+              if (v1 == v2) {
+                state &= 0b01;
+              } else {
+                state &= 0b10;
+              }
+              branch_state.PutBranchState(sig, state);
+              int state_copy = state;
+              int position = 1;
+              while (state_copy > 0) {
+                int bit = state_copy & 0b1;
+                if (bit == 1) {
+                  switch (position) {
+                    case 1:
+                      {
+                        double gap = v1 - v2;
+                        if (gap != 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 2:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                  }
+                }
+                state_copy >>= 1;
+                position++;
+              }
+            }
+            break;
+          case "I$>=":
+          case "IZ$>=":
+            {
+              Integer state = branch_state.GetBranchState(sig);
+              if (state == null) {
+                state = 0b11;
+              }
+              double prev_v1 = previous_vob.GetBranchValue1();
+              double prev_v2 = previous_vob.GetBranchValue2();
+
+              double v1 = vob.GetBranchValue1();
+              double v2 = vob.GetBranchValue2();
+              if (v1 >= v2) {
+                state &= 0b10;
+              } else {
+                state &= 0b01;
+              }
+              branch_state.PutBranchState(sig, state);
+              int state_copy = state;
+              int position = 1;
+              while (state_copy > 0) {
+                int bit = state_copy & 0b1;
+                if (bit == 1) {
+                  switch (position) {
+                    case 1:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap < 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 2:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                  }
+                }
+                state_copy >>= 1;
+                position++;
+              }
+            }
+            break;
+          case "I$<=":
+          case "IZ$<=":
+            {
+              {
+                Integer state = branch_state.GetBranchState(sig);
+                if (state == null) {
+                  state = 0b11;
+                }
+                double prev_v1 = previous_vob.GetBranchValue1();
+                double prev_v2 = previous_vob.GetBranchValue2();
+
+                double v1 = vob.GetBranchValue1();
+                double v2 = vob.GetBranchValue2();
+                if (v1 <= v2) {
+                  state &= 0b01;
+                } else {
+                  state &= 0b10;
+                }
+                branch_state.PutBranchState(sig, state);
+                int state_copy = state;
+                int position = 1;
+                while (state_copy > 0) {
+                  int bit = state_copy & 0b1;
+                  if (bit == 1) {
+                    switch (position) {
+                      case 1:
+                        {
+                          double gap = v1 - v2;
+                          double prev_gap = prev_v1 - prev_v2;
+                          if (prev_gap - gap < 0) {
+                            influence.put(sig, 1);
+                          }
+                        }
+                        break;
+                      case 2:
+                        {
+                          double gap = v1 - v2;
+                          double prev_gap = prev_v1 - prev_v2;
+                          if (prev_gap - gap > 0) {
+                            influence.put(sig, 1);
+                          }
+                        }
+                        break;
+                    }
+                  }
+                  state_copy >>= 1;
+                  position++;
+                }
+              }
+            }
+            break;
+          case "I$>":
+          case "IZ$>":
+            {
+              Integer state = branch_state.GetBranchState(sig);
+              if (state == null) {
+                state = 0b11;
+              }
+              double prev_v1 = previous_vob.GetBranchValue1();
+              double prev_v2 = previous_vob.GetBranchValue2();
+
+              double v1 = vob.GetBranchValue1();
+              double v2 = vob.GetBranchValue2();
+              if (v1 > v2) {
+                state &= 0b10;
+              } else {
+                state &= 0b01;
+              }
+              branch_state.PutBranchState(sig, state);
+              int state_copy = state;
+              int position = 1;
+              while (state_copy > 0) {
+                int bit = state_copy & 0b1;
+                if (bit == 1) {
+                  switch (position) {
+                    case 1:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap < 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 2:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                  }
+                }
+                state_copy >>= 1;
+                position++;
+              }
+            }
+            break;
+          case "I$<":
+          case "IZ$<":
+            {
+              Integer state = branch_state.GetBranchState(sig);
+              if (state == null) {
+                state = 0b11;
+              }
+              double prev_v1 = previous_vob.GetBranchValue1();
+              double prev_v2 = previous_vob.GetBranchValue2();
+
+              double v1 = vob.GetBranchValue1();
+              double v2 = vob.GetBranchValue2();
+              if (v1 < v2) {
+                state &= 0b01;
+              } else {
+                state &= 0b10;
+              }
+              branch_state.PutBranchState(sig, state);
+              int state_copy = state;
+              int position = 1;
+              while (state_copy > 0) {
+                int bit = state_copy & 0b1;
+                if (bit == 1) {
+                  switch (position) {
+                    case 1:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap < 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                    case 2:
+                      {
+                        double gap = v1 - v2;
+                        double prev_gap = prev_v1 - prev_v2;
+                        if (prev_gap - gap > 0) {
+                          influence.put(sig, 1);
+                        }
+                      }
+                      break;
+                  }
+                }
+                state_copy >>= 1;
+                position++;
+              }
+            }
+            break;
+        }
+      }
+    }
+    return influence;
   }
 }
