+++ test_agent_trace_reader/src/test/java/cn/yyx/labtask/test_agent_trace_reader/UseTraceReaderComparator.java
@@ -6,11 +6,10 @@ import java.util.Map;
 
 public class UseTraceReaderComparator {
   public static void main(String[] args) {
-    TraceReader tr = new TraceReader(new TestModel(), "seqBefore.txt", "seqAfter.txt");
+    TraceReader tr = new TraceReader("seqBefore.txt", "seqAfter.txt");
     Map<String, ValuesOfBranch> branchesInfo1 =
         tr.ReadFromTraceFile(System.getProperty("user.home") + "/trace1.txt");
     Map<String, ValuesOfBranch> branchesInfo2 =
         tr.ReadFromTraceFile(System.getProperty("user.home") + "/trace2.txt");
-
   }
 }
